restructure of the components folder to split by page, i think this is cleaner and makes sense for where the components are used
