+++ jobs/Frontend/movie-search/src/app/page.tsx
@@ -1,10 +1,10 @@
 import { Metadata } from "next";
-import MoviePagination from "@/components/MoviePagination/MoviePagination";
+import MoviePagination from "@/components/MovieHome/MoviePagination/MoviePagination";
 import { fetchMovies } from "@/data/FetchMovies";
-import MovieCard from "@/components/MovieCard/MovieCard";
-import MovieCardContainer from "@/components/MovieCard/MovieCardContainer";
-import MovieHeader from "@/components/MovieHeader/MovieHeader";
-import NoMoviesComponent from "@/components/NoMovies/NoMovies";
+import MovieCard from "@/components/MovieHome/MovieCard/MovieCard";
+import MovieCardContainer from "@/components/MovieHome/MovieCard/MovieCardContainer";
+import MovieHeader from "@/components/MovieHome/MovieHeader/MovieHeader";
+import NoMoviesComponent from "@/components/MovieHome/NoMovies/NoMovies";
 
 export const metadata: Metadata = {
   title: "MEWS Movie Search",
