+++ jobs/Frontend/movie-search/src/components/MovieDetails/MovieNavigation/MovieNavigation.tsx
