+++ jobs/Frontend/movie-search/src/components/MovieDetails/MovieNavigation/MovieNavigationStyledComponents.ts
