+++ jobs/Frontend/movie-search/src/components/MovieHome/MovieCard/MovieCard.tsx
