+++ jobs/Frontend/movie-search/src/components/MovieHome/MovieCard/MovieCardContainer.tsx
