+++ jobs/Frontend/movie-search/src/components/MovieHome/MovieCard/MovieCardStyledComponents.ts
